backend/requirements.txt
Django==5.0.7
channels==4.1.0
redis==5.0.7
channels_redis==4.2.0
daphne==4.1.2
flake8
Pillow
djangorestframework==3.15.2
django-cors-headers
djangorestframework-simplejwt==5.3.1
